pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aurum-edge"
version = "2.0.0"
description = "Motor de decisión ML para trading cuantitativo con validación walk-forward"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "LIA Engineering Solutions", email = "info@liaengineering.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Data manipulation
    "pandas>=2.1.0",
    "numpy>=1.24.0,<2.0.0",
    "pyarrow>=14.0.0",
    "fastparquet>=2023.10.0",
    
    # ML & Stats
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "optuna>=3.4.0",
    "scipy>=1.11.0",
    
    # Configuration & Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    
    # Logging & Monitoring
    "loguru>=0.7.2",
    "tqdm>=4.66.0",
    
    # Plotting (minimal for reports)
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    
    # Utils
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "tabulate>=0.9.0",
    
    # Telegram/Discord (optional)
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    "ipython>=8.18.0",
    "jupyter>=1.0.0",
]

mt5 = [
    "MetaTrader5>=5.0.45",  # Solo para fase 2
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "optuna.*",
    "xgboost.*",
    "sklearn.*",
    "scipy.*",
    "matplotlib.*",
    "seaborn.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src/aurum_edge",
    "--cov-report=term-missing",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
